requirements.txt
google-api-python-client >= 1.7.8
google-auth              >= 1.6.3
google-auth-httplib2     >= 0.0.3
SQLAlchemy               >= 1.2.16
paramiko                 >= 2.4.2
mysqlclient              >= 1.4.2